+++ CHANGES.md
@@ -5,6 +5,7 @@
 
 * Remove automatically resize
 * Change value conversion px to odt
+* Add extension at the end of the generated file
 
 1.2.22          (2019-11-25)
 ----------------------------